nova_flavors: reorganise into a package
