+++ operators/nova-flavors/nova_flavors/reconcile.py
@@ -5,25 +5,27 @@ import time
 from flavor_matcher.flavor_spec import FlavorSpec
 from watchdog.observers import Observer
 
-from flavor_synchronizer import FlavorSynchronizer
-from logger import setup_logger
-from spec_changed_handler import SpecChangedHandler
+from nova_flavors.flavor_synchronizer import FlavorSynchronizer
+from nova_flavors.logger import setup_logger
+from nova_flavors.spec_changed_handler import SpecChangedHandler
 
 loglevel = getattr(logging, os.getenv("NOVA_FLAVOR_MONITOR_LOGLEVEL", "info").upper())
 logging.getLogger().setLevel(loglevel)
 logger = setup_logger(__name__, level=loglevel)
 
-# nonprod vs prod
-FLAVORS_DIR = os.getenv("FLAVORS_DIR", "")
-if not os.path.isdir(FLAVORS_DIR):
-    raise ValueError(f"FLAVORS_DIR '{FLAVORS_DIR}' is not a directory")
+
+flavors_dir = ""
 
 
 def read_flavors():
-    return FlavorSpec.from_directory(FLAVORS_DIR)
+    return FlavorSpec.from_directory(flavors_dir)
 
 
 def main():
+    # nonprod vs prod
+    flavors_dir = os.getenv("FLAVORS_DIR", "")
+    if not os.path.isdir(flavors_dir):
+        raise ValueError(f"flavors_dir '{flavors_dir}' is not a directory")
     synchronizer = FlavorSynchronizer(
         username=os.environ.get("OS_USERNAME", ""),
         password=os.environ.get("OS_PASSWORD", ""),
@@ -34,8 +36,8 @@ def main():
 
     handler = SpecChangedHandler(synchronizer, read_flavors)
     observer = Observer()
-    observer.schedule(handler, FLAVORS_DIR, recursive=True)
-    logger.info(f"Watching for changes in {FLAVORS_DIR}")
+    observer.schedule(handler, flavors_dir, recursive=True)
+    logger.info(f"Watching for changes in {flavors_dir}")
     observer.start()
 
     try:
@@ -44,6 +46,7 @@ def main():
     finally:
         observer.stop()
         observer.join()
+
 
 if __name__ == "__main__":
     main()
